Offroad_Segmentation_Scripts/requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.7
tqdm>=4.66
fastapi>=0.115
uvicorn>=0.30
python-multipart>=0.0.9
